+++ CHANGELOG.md
@@ -3,8 +3,9 @@ CHANGELOG
 
 2.7.1.pre
 -----------------
+ * Update smart content element filters
  * Fix bugs for windows
- 
+
 2.7.0
 -----------------
  * It is ready